src/python/requirements.txt
requests==2.31.0
pymongo==4.5.0
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.8.5
asyncio==3.4.3
legiscan==1.0.0
# Added for compressor library
langchain==0.1.9
langchain-openai==0.0.5
openai==1.13.3
PyPDF2==3.0.1